+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/fluent/models/AvailabilitySetProperties.java
@@ -6,17 +6,13 @@ package com.azure.resourcemanager.compute.fluent.models;
 
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.SubResource;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.compute.models.InstanceViewStatus;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 /** The instance view of a resource. */
 @Fluent
 public final class AvailabilitySetProperties {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(AvailabilitySetProperties.class);
-
     /*
      * Update Domain count.
      */