+++ sdk/resourcemanager/azure-resourcemanager-compute/src/main/java/com/azure/resourcemanager/compute/models/ProximityPlacementGroupUpdate.java
@@ -5,15 +5,11 @@
 package com.azure.resourcemanager.compute.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.util.Map;
 
 /** Specifies information about the proximity placement group. */
 @Fluent
 public final class ProximityPlacementGroupUpdate extends UpdateResource {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(ProximityPlacementGroupUpdate.class);
-
     /** {@inheritDoc} */
     @Override
     public ProximityPlacementGroupUpdate withTags(Map<String, String> tags) {