+++ sdk/resourcemanager/azure-resourcemanager-containerservice/src/main/java/com/azure/resourcemanager/containerservice/implementation/OpenShiftManagedClustersClientImpl.java
@@ -30,7 +30,6 @@ import com.azure.core.management.exception.ManagementException;
 import com.azure.core.management.polling.PollResult;
 import com.azure.core.util.Context;
 import com.azure.core.util.FluxUtil;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.polling.PollerFlux;
 import com.azure.core.util.polling.SyncPoller;
 import com.azure.resourcemanager.containerservice.fluent.OpenShiftManagedClustersClient;
@@ -50,8 +49,6 @@ public final class OpenShiftManagedClustersClientImpl
         InnerSupportsListing<OpenShiftManagedClusterInner>,
         InnerSupportsDelete<Void>,
         OpenShiftManagedClustersClient {
-    private final ClientLogger logger = new ClientLogger(OpenShiftManagedClustersClientImpl.class);
-
     /** The proxy service used to perform REST calls. */
     private final OpenShiftManagedClustersService service;