+++ sdk/core/azure-core-http-okhttp/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 1.7.10 (2022-03-04)
+
+### Other Changes
+
+- Updated all `ClientLogger`s to be static constants instead of instance variables. ([#27339](https://github.com/Azure/azure-sdk-for-java/pull/27339))
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.25.0` to `1.26.0`.
+
 ## 1.7.9 (2022-02-04)
 
 ### Other Changes